% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_penguin.R
\name{classify_penguin}
\alias{classify_penguin}
\title{Classify a penguin}
\usage{
classify_penguin(bill_length, flipper_length)
}
\arguments{
\item{bill_length}{the length of bill in mm}

\item{flipper_length}{the length of flipper in mm}
}
\description{
Classify a penguin given the bill length and flipper length
}
