% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species.R
\name{get_species}
\alias{get_species}
\title{Returns the species column for Palmerpenguins}
\usage{
get_species()
}
\value{
dataframe of 1 column from the palrmer penguins dataframe
}
\description{
Returns the species column for Palmerpenguins
}
