% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_penguin.R
\name{classify_penguin}
\alias{classify_penguin}
\title{Classify a penguin}
\usage{
classify_penguin(bill_length, flipper_length)
}
\arguments{
\item{bill_length}{the length of a bill measured in mm}

\item{flipper_length}{the length of a flipper measured in mm}
}
\description{
Classifies a penguin given the bill length and flipper length.
}
\examples{
# example code
classify_penguin(10, 100)
}
