% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{get_species}
\alias{get_species}
\title{Returns the species column for Palmerpenguins}
\usage{
get_species()
}
\value{
dataframe of 1 column
}
\description{
Returns the species column for Palmerpenguins
}
\examples{
\dontrun{
get_species(not working)
}

}
